python-server/requirements.txt
tensorflow>=2.11
mediapipe==0.10.0
opencv-python==4.7.0.68
protobuf==3.20.3
grpcio-tools==1.48.2
